collection view basics - PMT #107908
I've separated out the logic for display the text popup vs. media popup.
I realize that this tool is probably not responsible for displaying the
media popup - we need to take advantage of the quick-edit and filtering
capabilitied of the existing Mediathread collection browser. When I
started this branch, my thoughts were that it would be most
straightforward to expose some JSON from the back-end to build this
view, but I now realize there's too much functionality to rebuild, and
we want the familiar Mediathread browser here. So, after some
investigation it looks like we can pretty easily call external
JavaScript from within this React app. I still have to investigate how
the collection browser works on the Mediathread side.