Only update video players on mouseUp - PMT #109139
This improves the behavior a bit by not constantly seeking as you drag
the playhead. I looked at this code here to see how the author of
react-player uses this library:

* https://github.com/CookPete/react-player/blob/master/src/demo/App.js

Which runs this player:

* http://rplayr.com/user/evilnight/m/truemusic

Note that the problem of the playhead "jumping" to the wrong position
remains, but now it will be much easier to debug that bug.