Don't setDirty in onSpineDuration - PMT #109410
This was causing the dirty state to be set whenever a spine video loaded
its duration, causing the referenced bug. The intention here was that
the dirty state should be set whenever track elements could have been
removed, which could occur with the removeOutOfBoundsElements() calls in
this function. But this case is already handled in
onOutOfBoundsConfirmClick().