Use setState callback when setting state based on itself
When a state update uses part of its own value as part of the new value,
react has a special mechanism for this:

From the setState docs:
> It's possible to pass a function with the signature function(state,
> props) => newState. This enqueues an atomic update that consults the
> previous value of state and props before setting any values.
https://facebook.github.io/react/docs/react-component.html#setstate

I haven't seen any problems caused by the existing code, but I figured
it can't hurt to program defensively here.

I've also updated the state update code in `editActiveMediaTrackElement`
- this function was calling `_.extend({}, self.state);`, which is
unnecessary because then all state attributes are updated instead of
just the one we care about, which is `mediaTrack` in this case.