TrackElementManager Enter key fix - PMT #109459
Don't trigger the 'Edit Selection' button when pressing enter in the
TrackElementManager form. Since we're not using any traditional form
events here, I've just changed the form to a div.

In fact, when a timecode input is selected, pressing enter should
trigger the 'change' event. I'll do that in a separate PR.