Include babel-polyfill for old browser support - PMT #109527
So I'm not able to reproduce this bug in Safari 10, and Zarina's using
Safari 8. Searching around for the error, it looks like it's recommended
to include babel-polyfill to our build process for old browsers that
don't have new globals like Promise defined.
- https://github.com/babel/babelify/issues/22

We already had this in package.json but it wasn't being used. I've
activated it in our webpack config with these instructions:
http://babeljs.io/docs/usage/polyfill/#usage-in-node-browserify-webpack