Always cue players that are about to play - PMT #109528
This adds some code that explicitly seeks the player to the right place
when it's about to play. From what I've seen, seeking on a media element
while the player is visible is working well, and this should address at
least some of the problems where rewinding to before the element and
playing would play the wrong selection.