Add rewind button - PMT #110065
Because the player doesn't get immediately paused when we set `playing`
to `false`, the video continues to play for a short time period after
clicking the rewind button. So, it was necessary to add the
`setTimeout()` call to catch this case. This results in jumpy playhead
behavior, but it does work.